Fix crash on admin page
There was a crash on the admin page due to wrongly localized strings in
the HeroComponent. The strings where not loaded when going to the /admin
route, I couldn't figure out what was wrong and ended up writing my own
i18n library to get better control over the localization. By adding
better error handling it was easy to spot that some of the strings in
HeroComponent where in the wrong group.

Moving them to the right group solved the issue.