Change from nested slug url to just single slug
We want to simplify the code related to blog posts. We do this in two
ways, we wrap supabase in a api endpoint. This allows us to not expose
supabase to the client.

We also move from [...slugs] to [slug] as the route parameter for
accessing blogposts. This makes the code much less complex.