Migrate UI to use @skeleton as UI framework
Initial tests show that we get a 99% performance score when using
skeleton. That is good enough so we can proceed with migrating the old
UI. It also gives us the foundation for theming the website and provides
a lot of UI components and styles that we want. With the old code we
would have needed to reimplement @rgossiaux/svelte-headlessui to get
proper performance from the app. Also remove any front end deps that we
won't be using like google analytics and sentry.