Make it possible for vite to generate types for .gql files
We switched from houdini to using urql since it is more used and easier
to use if we don't plan on doing anything magic. This allows us to
easily use graphql on the serverside. The goal is to never expose
https://jardoole.xyz to the client but handle all communication to it on
the server. Hopefully this will make it possible to build the site
statically removing the need to rely on https://jardoole.xyz in
production, since it will only be needed at build time. If possible we
will be triggereing builds when articles or Terapheuts are updated
leaving us with a always up to date static site.

If it doesn't work we will look at caching.