Stop using svelte-heroicons
When we upgraded to svelte v4 only sveltekit was updated, apparently
svelte-heroicons prevented us to upgrade svelte. svelte-heroicons is not
upgraded to use svelte v4. Instead of waiting for them to update we
simply move the 4 icons we where using them for directly to our code and
remove the dependency.