Add banner to / page
We want to be able to add news to a easy to see place on the website.
Therefore we create a banner that we place under the header on the /
route. It can easily be changed from the cms system and no banner will
be shown if removed from the cms system.