Redirect to ww1.clinicbuddy.com for booking
clinicbuddy has added cloudflare for ddos protection and we don't have
an exception. When we try to load our booking page it thinks that our
users are ddosing them and thus they can't book. To mitigate this we
redirect to ww1.clinicbuddy.com instead of keeping the users on our
site.